requirements.txt
pydantic>=2.0
pyyaml>=6.0
pillow>=10.0
numpy>=1.24.0